Reorganized file hierarchy. Added docker-compose.yml to organize services. Updated dockerfile to run bot unbuffered. Fixed move command.
