--- Dockerfile
@@ -1,9 +1,0 @@
-FROM python:3.11-alpine
-
-WORKDIR /app
-
-COPY requirements.txt .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-CMD ["python", "main.py"]
