+++ bot/.dockerignore
@@ -1,0 +1,6 @@
+**/__pycache__/
+**/*.pyc
+
+Dockerfile
+.dockerignore
+README.md
