+++ bot/media/discordlogogif.gif
