+++ bot/requirements.txt
@@ -1,0 +1,3 @@
+discord.py==2.4.0
+pymongo==4.10.1
+python-dotenv==1.0.1
