--- requirements.txt
@@ -1,2 +1,0 @@
-discord.py==2.4.0
-pymongo==4.10.1
